c/fstream: ifstm/ofstm are now variable-size
The user should define their own struct whose only
two elements are an ifstm/ofstm (the header)
immediately followed by an uint8_t array of the
desired size (2*page_size for ifstm, or buf_size
for ofstm).

This lets the user define custom "subclasses" of
ifstm/ofstm with different buffer sizes, and all
of them can be allocated on the stack.